[AssetMapper] Improve import statements extraction
This PR introduce a tiny JS parser responsible to split the code into distinct parts:
* default (code)
* comments
* enclosed strings

By parsing step by step, only analysing what's necessary in parallel of the preg_replace_callback, it may ease the Compiler choice to keep or not an import statement.

This also allow us to use a lighter regexp, avoiding any JIT trouble.

It was just a game/test at first, but as it seem to work really well, let's open the discussion :)

# Conflicts:
#	src/Symfony/Component/AssetMapper/Compiler/JavaScriptImportPathCompiler.php